BugFix: Processing Multiple Files Leaves Tokens in the AlphaSniff Instance
A single instance of the AlphabeticalUseSniff is used to process multiple files. This class has an instance variable that tracked the last seen token to be used in a comparison with the next token. When the file that's being inspected changes from one to another, this variable is not cleared and the first matching token of the new file is compared against the last found token of the previous file. This is obviously incorrect.

Converting the instance variable to a per-file map resolves this issue. An alternative would be to clear out and reinitialize the array and and have an instance variable tracking the file - although more efficient, would be clunkier.